Rework FPlayers and remove remnants of steamchart
