+++ src/html/options.html
@@ -187,7 +187,7 @@
                     <div data-dynamic="showoc"></div>
                     <div data-dynamic="showhltb"></div>
                     <div data-dynamic="showwsgf"></div>
-                    <div data-dynamic="show_steamchart_info"></div>
+                    <div data-dynamic="show_players_info"></div>
                     <div data-dynamic="showastatslink"></div>
                     <div data-dynamic="showitadlinks"></div>
                     <div data-dynamic="showsteamdb"></div>
